Restructure, Dependency upgrades, Service Worker
