--- .babelrc
@@ -1,5 +1,0 @@
-{
-    "presets": [
-      "@babel/react", "@babel/typescript", ["@babel/env", { "modules": false }]
-    ]
-  }
