+++ package.json
@@ -3,75 +3,76 @@
   "version": "1.0.0",
   "description": "A conference template system built in react, redux, and firebase. 44x better than v2",
   "scripts": {
-    "clean": "if exist build rmdir /S /Q build",
-    "start": "npm run build && cross-env NODE_ENV='development' webpack-dev-server --content-base src/ --inline --hot --history-api-fallback",
-    "build": "npm run clean && cross-env NODE_ENV='development' webpack --verbose --colors",
-    "build-prod": "npm run clean && cross-env NODE_ENV='production' webpack -p --verbose --colors",
-    "build-watch": "npm run clean && cross-env NODE_ENV='development' webpack --watch --verbose --colors"
+    "clean": "rm -rf build",
+    "build": "npm run clean && NODE_ENV='development' webpack --verbose --colors",
+    "build-prod": "npm run clean && NODE_ENV='production' webpack -p --verbose --colors",
+    "build-watch": "NODE_ENV='development' webpack --watch --verbose --display --display-error-details --colors"
   },
   "author": "Jeremy Shore",
   "license": "MIT",
   "dependencies": {
-    "@firebase/app": "^0.3.4",
-    "@firebase/auth": "^0.7.7",
-    "@firebase/firestore": "^0.8.2",
-    "@firebase/storage": "^0.2.3",
-    "@material-ui/core": "^3.1.0",
+    "@date-io/date-fns": "0.0.2",
+    "@firebase/app": "^0.3.5",
+    "@firebase/auth": "^0.7.9",
+    "@firebase/firestore": "^0.8.8",
+    "@firebase/storage": "^0.2.4",
+    "@material-ui/core": "^3.5.1",
     "@material-ui/icons": "^3.0.1",
     "classnames": "^2.2.6",
-    "date-fns": "^2.0.0-alpha.16",
+    "date-fns": "^2.0.0-alpha.21",
     "isomorphic-fetch": "^2.2.1",
-    "material-ui-pickers": "^1.0.0-rc.17",
-    "react": "^16.5.2",
+    "material-ui-pickers": "^2.0.2",
+    "react": "^16.6.3",
     "react-async-script": "^1.0.0",
-    "react-dom": "^16.5.2",
-    "react-dropzone": "^6.2.1",
-    "react-pose": "^3.3.4",
-    "react-redux": "^5.0.7",
+    "react-dom": "^16.6.3",
+    "react-dropzone": "^7.0.1",
+    "react-pose": "^4.0.2",
+    "react-redux": "^5.1.1",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "redux": "^4.0.0",
-    "redux-actions": "^2.6.1",
-    "redux-saga": "^0.16.0",
-    "reselect": "^3.0.1",
-    "showdown": "^1.8.6"
+    "redux": "^4.0.1",
+    "redux-actions": "^2.6.4",
+    "redux-saga": "^0.16.2",
+    "reselect": "^4.0.0",
+    "showdown": "^1.9.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.0",
-    "@babel/preset-env": "^7.1.0",
+    "@babel/core": "^7.1.6",
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/preset-env": "^7.1.6",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.1.0",
-    "@types/googlemaps": "^3.30.13",
-    "@types/node": "^10.10.1",
-    "@types/react": "^16.4.14",
-    "@types/react-dom": "^16.0.7",
+    "@types/googlemaps": "^3.30.16",
+    "@types/node": "^10.12.9",
+    "@types/react": "^16.7.6",
+    "@types/react-dom": "^16.0.9",
     "@types/react-redux": "^6.0.9",
-    "@types/react-router": "^4.0.31",
+    "@types/react-router": "^4.4.1",
     "@types/react-router-dom": "^4.3.1",
-    "@types/redux-actions": "^2.3.0",
-    "@types/webpack": "^4.4.13",
-    "babel-loader": "^8.0.2",
+    "@types/redux-actions": "^2.3.1",
+    "@types/webpack": "^4.4.19",
+    "babel-loader": "^8.0.4",
     "babel-polyfill": "^6.26.0",
-    "cross-env": "^5.2.0",
-    "css-loader": "^1.0.0",
+    "css-loader": "^1.0.1",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "favicons-webpack-plugin": "^0.0.9",
     "file-loader": "^2.0.0",
     "html-webpack-plugin": "^3.2.0",
-    "node-sass": "^4.9.3",
+    "image-webpack-loader": "^4.5.0",
+    "node-sass": "^4.10.0",
     "path": "^0.12.7",
     "react-svg-loader": "^2.1.0",
     "resolve-url-loader": "^3.0.0",
     "sass-loader": "^7.1.0",
     "source-map-loader": "^0.2.4",
-    "style-loader": "^0.23.0",
-    "ts-loader": "^5.1.1",
+    "style-loader": "^0.23.1",
+    "ts-loader": "^5.3.0",
     "tslib": "^1.9.3",
     "tslint": "^5.11.0",
     "tslint-react": "^3.6.0",
-    "typescript": "^3.0.3",
-    "webpack": "^4.19.1",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.8"
+    "typescript": "^3.1.6",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.26.0",
+    "webpack-cli": "^3.1.2"
   }
 }
