+++ src/components/dialogs/Sponsor/Sponsor.scss
@@ -1,0 +1,56 @@
+.drag-drop-logo {
+    width: 35%;
+    border-width: 2px;
+    border-color: rgb(102, 102, 102);
+    border-style: dashed;
+    border-radius: 15px;
+    max-height: 95px;
+    margin: 16px;
+    min-height: 110px;
+    display: flex;
+    align-items: center;
+    transition: all .5s ease;
+
+    &:hover {
+        background-color: rgba(150,150,150,0.20);
+        cursor: pointer;
+    }
+
+    > p {
+        margin-left: auto;
+        margin-right: auto;
+        opacity: .5;
+    }
+}
+
+.drag-drop-preview {
+    width: 35%;
+    margin: 16px;
+    display: flex;
+
+    > div {
+        margin: auto 0;
+    }
+
+    img {
+        display: block;
+        margin: 0 auto;
+        max-width: 80%;
+        max-height: 90px;
+        width: auto;
+    }
+}
+
+.create-sponsor-form {
+    display: flex;
+    flex-direction: row;
+
+    .sponsor-details {
+        flex: 1;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        margin-top: auto;
+        margin-bottom: auto;
+    }
+}
