+++ src/components/sections/Sponsors/Sponsors.scss
@@ -1,0 +1,36 @@
+.logo-list {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: flex-start;
+
+    li {
+        width: 33.333333%;
+        padding: 20px 5px;
+        transition: all .5s ease;
+        border-radius: 8px;
+
+        &:hover {
+            -webkit-box-shadow: 0.4rem 0.4rem 2rem rgba(0,0,0,0.7);
+            box-shadow: 0.4rem 0.4rem 2rem rgba(0,0,0,.7);
+        }
+    }
+
+    img {
+        display: block;
+        margin: 0 auto;
+        max-width: 80%;
+        max-height: 90px;
+        width: auto;
+    }
+}
+
+@media (min-width: 544px) {
+
+    .logo-list li {
+        width: calc(100% / 6);
+        flex-grow: 1;
+        max-width: calc(100% / 6);
+    }
+
+}
