+++ src/config/delorean.config.js.template
