+++ src/ducks/current.ts
@@ -7,6 +7,16 @@ import Configuration from '../models/config';
 import { FirebaseApp } from '@firebase/app-types';
 import { User } from '@firebase/auth-types';
 import Sponsor from '../models/sponsor';
+import { ApplicationState } from '..';
+import { FirebaseFirestore } from '@firebase/firestore-types';
+
+export const getFirebaseApp = (state: ApplicationState): FirebaseApp => state.current.firebase;
+export const getFirestore = (state: ApplicationState): FirebaseFirestore => state.current.firebase.firestore();
+export const getUser = (state: ApplicationState) => state.current.user;
+export const getUserProfile = (state: ApplicationState) => state.current.profile;
+export const getCurrentConfig = (state: ApplicationState) => state.current.config;
+export const getIsEditMode = (state: ApplicationState) => state.current.isEditMode;
+export const getSponsors = (state: ApplicationState) => state.current.sponsors;
 
 export enum CurrentTypes {
     SET_USER = 'SET_USER',
