--- src/selectors/config.ts
@@ -1,3 +1,0 @@
-import { ApplicationState } from '..';
-
-export const isConfigDialogOpen = (state: ApplicationState) => state.config.isOpen;
