+++ src/stylesheets/globals/_tickets.scss
