+++ src/stylesheets/main.scss
@@ -1,0 +1,14 @@
+@import './globals/variables';
+@import './globals/mixins';
+@import './globals/keyframes';
+@import './globals/placeholders';
+@import './globals/media';
+@import './globals/global';
+@import './globals/tickets';
+
+@import '../components/controls/Header/Header.scss';
+@import '../components/controls/Footer/Footer.scss';
+@import '../components/controls/EditOverlay/EditOverlay.scss';
+@import '../components/dialogs/Sponsor/Sponsor.scss';
+@import '../components/pages/Home/Home.scss';
+@import '../components/sections/Sponsors/Sponsors.scss'
