--- stylesheets/components/_editoverlay.scss
@@ -1,26 +1,0 @@
-.overlay {
-    pointer-events: none;
-    background: none;
-    overflow: visible;
-    position: fixed;
-    z-index: 1000;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-
-    .fab-container {
-        display: flex;
-        height: 100%;
-        background: none;
-        overflow: visible;
-        pointer-events: none;
-    }
-
-    .edit-fab {
-        margin-left: auto;
-        margin-top: auto;
-        margin-bottom: 32px;
-        pointer-events: all;
-    }
-}
