--- stylesheets/components/_sponsors.scss
@@ -1,93 +1,0 @@
-.logo-list {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: flex-start;
-
-    li {
-        width: 33.333333%;
-        padding: 20px 5px;
-        transition: all .5s ease;
-        border-radius: 8px;
-
-        &:hover {
-            -webkit-box-shadow: 0.4rem 0.4rem 2rem rgba(0,0,0,0.7);
-            box-shadow: 0.4rem 0.4rem 2rem rgba(0,0,0,.7);
-        }
-    }
-
-    img {
-        display: block;
-        margin: 0 auto;
-        max-width: 80%;
-        max-height: 90px;
-        width: auto;
-    }
-}
-
-.drag-drop-logo {
-    width: 35%;
-    border-width: 2px;
-    border-color: rgb(102, 102, 102);
-    border-style: dashed;
-    border-radius: 15px;
-    max-height: 95px;
-    margin: 16px;
-    min-height: 110px;
-    display: flex;
-    align-items: center;
-    transition: all .5s ease;
-
-    &:hover {
-        background-color: rgba(150,150,150,0.20);
-        cursor: pointer;
-    }
-
-    > p {
-        margin-left: auto;
-        margin-right: auto;
-        opacity: .5;
-    }
-}
-
-.drag-drop-preview {
-    width: 35%;
-    margin: 16px;
-    display: flex;
-
-    > div {
-        margin: auto 0;
-    }
-
-    img {
-        display: block;
-        margin: 0 auto;
-        max-width: 80%;
-        max-height: 90px;
-        width: auto;
-    }
-}
-
-.create-sponsor-form {
-    display: flex;
-    flex-direction: row;
-
-    .sponsor-details {
-        flex: 1;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        margin-top: auto;
-        margin-bottom: auto;
-    }
-}
-
-@media (min-width: 544px) {
-
-    .logo-list li {
-        width: calc(100% / 6);
-        flex-grow: 1;
-        max-width: calc(100% / 6);
-    }
-
-}
