--- stylesheets/main.scss
@@ -1,13 +1,0 @@
-@import './globals/_variables.scss';
-@import './globals/_mixins.scss';
-@import './globals/_keyframes.scss';
-@import './globals/_placeholders.scss';
-@import './globals/_media.scss';
-@import './globals/_global.scss';
-
-@import './components/home';
-@import './components/header';
-@import './components/footer';
-@import './components/tickets';
-@import './components/editoverlay';
-@import './components/sponsors';
