--- tsconfig.json
@@ -1,30 +1,0 @@
-{
-    "compilerOptions": {
-      "outDir": "build/dist",
-      "module": "commonjs",
-      "target": "es5",
-      "lib": ["es6", "dom"],
-      "sourceMap": true,
-      "jsx": "react",
-      "moduleResolution": "node",
-      "declaration": false,
-      "importHelpers": true,
-      "removeComments": true,
-      "preserveConstEnums": true,
-      "strictNullChecks": false,
-      "typeRoots": [
-        "node_modules/@types",
-        "node_modules"
-      ],
-      "types": [
-        "googlemaps", "node", "react", "react-dom", "react-redux",
-        "react-router", "react-router-dom", "redux-actions"
-      ]
-    },
-    "include": [
-      "src",
-      // "node_modules/**/*.d.ts",
-      // "node_modules/@types/**/*.d.ts",
-      "typings.d.ts"
-    ]
-}
